Fix Makefile
Code compiled correctly before only when no functions nor subroutines
from the subroutines module were called, as the subroutines.mod file was
correctly created. However, it failed once a subroutine from the module
was called due to the subroutines.o not being created nor added to the
compilation of the main program.

This commit fixes those issues and the code can now be compiled without
issues and the subrouitnes can be used.