Update positions output file formatting
Positions output file will be structured as follows from now on
(considering a 3D system with cartesian coordinates):
time, x, y, z

The write_positions_xyz subroutine has been updated to reflect this
change. On each time iteration, results will be appended.

An example on how it must be used was added in the main file.