integrators.f90: Remove variable allocation
Remove allocation of variables that are already allocated.
this subroutines are executed in a loop, resulting in a memory error if
the allocation is done repeatedly (because the variable is already
allocated). Allocation should be done outside these subroutines and
before executing them.

This is prove that whoever did this part didn't test anything.