Remake binning plotting script
The binning plotting script didn't follow the same standards as the rest
of plots. This commits fixes that.

The binning plotting script needs to be called with the file containing
the data.

The script now saves the figures as a PDF instead of displaying it.

The script is added to plot.sh