geometry.f90: Use part_num defined in global_vars
Instead of determining part_num from the positions array, use the value
specified in the input_parameters.in file that is then stored as defined
in the global_vars module.

This will avoid having to determine the size of the positions array
everytime PBC are applied, improving the program's performance.