(parallel): Create 'counts' and 'displs' in main (vdw_gas)
The 'counts' and 'displs' arrays, necessary for 'mpi_allgatherv'
operations are just created once in the main program, instead of being
created in all the subroutines.

As they are allocatable arrays, they need to be passed as arguments to
the subroutines that need them in order to avoid memory access problems.