vdw_gas (parallel): Fix output file opening and closing
As only rank 0 (the master/root process) will be writing to the output
files, the opening and closing of said files must only be made by said
process.

Although the files being opened and closed by all the process doesn't
affect the results, as the data is correclty written (only rank 0
modifies the files), it can affect the overall program performance.