io (parallel): Update positions output file writing
Update the 'write_positions_xyz' subroutine.

Opening and closing the file on each iteration is very resource
intensive and this slows a lot the program. This becomes much more
noticeable in the parallel implementation of the program.

This commit solves that issue, by only opening and closing the file once
in the main program and using the 'write_positions_xyz' subroutine only
for writing the file contents.